+++ pyslvs/optimization/__init__.py
@@ -7,6 +7,6 @@ __copyright__ = "Copyright (C) 2016-2021"
 __license__ = "AGPL"
 __email__ = "pyslvs@gmail.com"
 
-from .planar import (
+from .f_planar import (
     FMatch, norm_path, curvature, derivative, path_signature, cross_correlation,
 )