+++ changelog.md
@@ -1,6 +1,6 @@
 # Version 21.02
 
-+ None.
++ Add function callback counter for `FMatch`.
 
 # Version 21.01