+++ changelog.md
@@ -1,3 +1,7 @@
+# Version 21.02
+
++ Add function callback counter for `FMatch`.
+
 # Version 21.01
 
 + Add examples: